requirements.txt
fastmcp>=0.1.7
requests>=2.31.0
python-dotenv>=1.0.0 